+++ common/src/main/java/com/iamkaf/liteminer/config/LiteminerClientConfig.java
@@ -5,9 +5,12 @@ import net.neoforged.neoforge.common.ModConfigSpec;
 public final class LiteminerClientConfig {
     public final ModConfigSpec.ConfigValue<KeyMode> keyMode;
     public final ModConfigSpec.ConfigValue<Boolean> showHUD;
+    public final ModConfigSpec.ConfigValue<Double> hud_scale;
 
     public LiteminerClientConfig(ModConfigSpec.Builder builder) {
         keyMode = builder.translation("liteminer.config.key_mode").defineEnum("key_mode", KeyMode.HOLD);
         showHUD = builder.translation("liteminer.config.show_hud").define("show_hud", true);
+        hud_scale =
+                builder.translation("liteminer.config.hud_scale").defineInRange("hud_scale", 1d, 0.5d, 2d);
     }
 }